requirements_web.txt
streamlit>=1.28.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
moviepy>=1.0.3
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0 